# optimize default case for to_hex method